fix: rename shopName to shop; change Integer to Long

+++ backend/src/main/java/it/compare/backend/product/response/ProductListResponse.java
@@ -15,6 +15,6 @@ public class ProductListResponse {
     private String mainImageUrl;
     private BigDecimal lowestCurrentPrice;
     private String lowestPriceShop;
-    private Integer offerCount;
+    private Long offerCount;
     private Boolean isAvailable;
 }
